third_party/blink/web_tests/platform/linux-chrome/external/wpt/visual-viewport/resize-event-order-expected.txt

This is a testharness.js-based test.
[FAIL] Popup: DOMWindow resize fired before VisualViewport.
  assert_equals: expected "window-attribute,window-addEventListener,visualViewport-attribute,visualViewport-addEventListener" but got ""
Harness: the test ran to completion.
